--- Dockerfile.modern
@@ -1,28 +1,0 @@
-# Modern Dockerfile for FastMCP 2.0
-FROM python:3.12-slim
-
-# Set working directory
-WORKDIR /app
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y \
-    --no-install-recommends \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements and install Python dependencies
-COPY requirements_modern.txt .
-RUN pip install --no-cache-dir -r requirements_modern.txt
-
-# Copy source code
-COPY src/ ./src/
-
-# Set environment variables
-ENV PYTHONPATH=/app
-ENV PYTHONUNBUFFERED=1
-
-# Health check
-HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
-    CMD python -c "import sys; sys.exit(0)"
-
-# Default command
-CMD ["python", "-m", "src.main_modern"]
