+++ FINAL_CLEANUP_STATUS.md
@@ -1,0 +1,56 @@
+# 🎯 FINAL CLEANUP STATUS - IMPLEMENTATION COMPLETE
+
+## ✅ Current File Structure (CORRECT)
+
+```
+src/
+├── main.py          ✅ Modern FastMCP 2.0 implementation (ACTIVE)
+├── main_legacy.py   ✅ Legacy implementation (BACKUP)
+├── utils.py         ✅ Greptile API client (unchanged)
+└── __init__.py      ✅ Package initialization
+
+Root Directory:
+├── requirements.txt      ✅ Modern dependencies (fastmcp>=2.10.0)
+├── requirements_legacy.txt ✅ Legacy dependencies (backup)
+├── Dockerfile           ✅ Modern container build
+├── Dockerfile.legacy    ✅ Legacy container (backup)
+└── smithery.yaml        ✅ Points to src.main (correct)
+```
+
+## ✅ No Issues Found
+
+The file structure is actually **CORRECT AS-IS**:
+
+1. **Modern implementation** is active in `src/main.py`
+2. **Legacy backup** is preserved as `src/main_legacy.py`
+3. **No duplicate** `src/main_modern.py` exists (good!)
+4. **All production files** use the modern implementation
+
+## 📋 Optional Future Cleanup
+
+After confirming stable production operation, you may optionally:
+
+1. **Delete legacy backups**:
+   ```bash
+   rm src/main_legacy.py
+   rm requirements_legacy.txt
+   rm Dockerfile.legacy
+   ```
+
+2. **Archive documentation**:
+   ```bash
+   mkdir archive
+   mv modernization_plan.md archive/
+   mv MIGRATION_CHECKLIST.md archive/
+   ```
+
+## 🎯 FINAL STATUS: READY TO CLOSE OUT
+
+**✅ The implementation is COMPLETE and CORRECT**
+
+- Modern FastMCP 2.0 is active in production files
+- Legacy files are properly backed up
+- No duplicate or conflicting files exist
+- Ready for immediate Smithery deployment
+
+**No further action required!** 🎉
