+++ requirements_legacy.txt
@@ -1,0 +1,3 @@
+httpx>=0.27.0
+mcp[cli]>=1.3.0
+python-dotenv>=1.0.0
