--- requirements_modern.txt
@@ -1,11 +1,0 @@
-# Modern FastMCP 2.0 Dependencies
-fastmcp>=2.10.0
-
-# HTTP client for Greptile API
-httpx>=0.27.0
-
-# Environment variable management
-python-dotenv>=1.0.0
-
-# Optional: Enhanced async support
-uvloop>=0.19.0; sys_platform != "win32"
